v13.2.5
For the conversion from "XML" to "DataModel" and vice versa, you
previously needed a corresponding "Transformer" class for each
"DataModel" class.
In this version "Transformer" classes are no longer required for the
conversion.
For the automatic conversion, Java's "Reflaction API" was used in this
version.

The automatic conversion algorithms are not quite finished yet.