v13.3.0
In this version I separated the "FinderSupport" and "SorterSupport"
algorithms from the "StoreSupper" class and created new "Support"
classes called "FinderSupport" and "SorterSupport".
And I also took the generation of "MagicLamp" and "Genie" from the
"StoreSupport" class and built the necessary directly into the
corresponding classes.
And finally I removed the "StoreSupport" class again.